Add files via upload
An analysis of the battles fought in the 'Westros' from the book - 'Songs of Ice and Fire' by George R. R. Martin which is also one of the highest viewed series, better known as 'Game of Thrones'. The analysis shows the performance of different kings both while attacking and defending. It shows which king had better win percentage over other kings. It also shows the battlegrounds where most battles were fought and the types of battles that were fought.