Add files via upload
The analysis of IPL Data from 2008 season to 2017 season. It shows the winners of each season and the number of matches won by each team for the 2017 season. It shows the ratio of matches won over toss wins by each team. It also depicts the number of matches won when a team won the toss and decided to field first or bat first.